extracted all extensions to modules dir - wrapping them not inside the dir.
commit-id:6aedbca4
