+++ crates/sierra/src/extensions/modules/mod.rs
@@ -1,0 +1,19 @@
+use super::SpecializationError;
+use crate::ids::ConcreteTypeId;
+use crate::program::GenericArg;
+
+pub mod felt;
+pub mod function_call;
+pub mod gas;
+pub mod integer;
+pub mod mem;
+pub mod non_zero;
+pub mod unconditional_jump;
+
+/// Helper for extracting the type from the template arguments.
+fn as_single_type(args: &[GenericArg]) -> Result<ConcreteTypeId, SpecializationError> {
+    match args {
+        [GenericArg::Type(ty)] => Ok(ty.clone()),
+        _ => Err(SpecializationError::UnsupportedGenericArg),
+    }
+}
